improve and simplify handling of syzygy_search_path
There's a bug in the Fathom Syzygy library that reads past the end of
search path strings if they end with a separator character.  valgrind
complained about this.  Fix it here in this code by using boost's
string join function instead of the painfully complicated way of doing
it by hand.